+++ src/olmo_core/distributed/utils.py
@@ -38,10 +38,11 @@ def init_distributed(backend: str = "nccl", timeout: timedelta = timedelta(minut
 
     # Set host-specific env var defaults.
     if _running_in_beaker():
+        multi_node = int(os.environ.get(OLMO_NUM_NODES_ENV_VAR, "1")) > 1
         # See https://beaker-docs.apps.allenai.org/experiments/distributed-training.html
         if "jupiter" in get_node_hostname():
             set_env_var("NCCL_IB_HCA", "^=mlx5_bond_0")
-            if int(os.environ.get(OLMO_NUM_NODES_ENV_VAR, "1")) > 1:
+            if multi_node:
                 # Only for multi-node
                 set_env_var("NCCL_SOCKET_IFNAME", "ib")
         elif "pluto" in get_node_hostname():
@@ -68,11 +69,25 @@ def init_distributed(backend: str = "nccl", timeout: timedelta = timedelta(minut
                 "NCCL_FASTRAK_IFNAME",
                 "enp6s0,enp7s0,enp13s0,enp14s0,enp134s0,enp135s0,enp141s0,enp142s0",
             )
-            set_env_var("NCCL_SOCKET_IFNAME", "enp0s12")
             set_env_var("NCCL_USE_SNAP", "1")
             set_env_var("NCCL_FASTRAK_USE_LLCM", "1")
-
-    validate_env_vars()
+            set_env_var("NCCL_FASTRAK_LLCM_DEVICE_DIRECTORY", "/dev/aperture_devices")
+            # NOTE: This path var must be set prior to launching Python
+            #  set_env_var(
+            #      "LD_LIBRARY_PATH",
+            #      "/var/lib/tcpxo/lib64:" + os.environ.get("LD_LIBRARY_PATH", ""),
+            #      override=True,
+            #  )
+            set_env_var("NCCL_TUNER_PLUGIN", "libnccl-tuner.so")
+            set_env_var(
+                "NCCL_TUNER_CONFIG_PATH", "/var/lib/tcpxo/lib64/a3plus_tuner_config.textproto"
+            )
+            set_env_var(
+                "NCCL_SHIMNET_GUEST_CONFIG_CHECKER_CONFIG_FILE",
+                "/var/lib/tcpxo/lib64/a3plus_guest_config.textproto",
+            )
+            if multi_node:
+                set_env_var("NCCL_SOCKET_IFNAME", "enp0s12")
 
     if backend_supports_cuda(backend):
         # Set CUDA device.
@@ -82,6 +97,8 @@ def init_distributed(backend: str = "nccl", timeout: timedelta = timedelta(minut
         torch.cuda.set_device(device)
 
     dist.init_process_group(backend, timeout=timeout)
+
+    validate_env_vars()
 
 
 def validate_env_vars():
@@ -94,12 +111,12 @@ def validate_env_vars():
     if OLMO_LOCAL_RANK_ENV_VAR not in os.environ:
         raise OLMoEnvironmentError(f"Missing env var '{OLMO_LOCAL_RANK_ENV_VAR}'")
 
-    if (
-        os.environ.get(OLMO_SHARED_FS_ENV_VAR) != "1"
-        and os.environ.get(OLMO_FS_LOCAL_RANK_ENV_VAR) is None
+    if os.environ.get(OLMO_SHARED_FS_ENV_VAR) != "1" and (
+        os.environ.get(OLMO_FS_LOCAL_RANK_ENV_VAR) is None
+        and os.environ.get(OLMO_LOCAL_RANK_ENV_VAR) is None
     ):
         raise OLMoEnvironmentError(
-            f"Missing env var '{OLMO_FS_LOCAL_RANK_ENV_VAR}' for non-shared filesystem. "
+            f"Missing env var '{OLMO_FS_LOCAL_RANK_ENV_VAR}'/'{OLMO_LOCAL_RANK_ENV_VAR}' for non-shared filesystem. "
             f"If this is a shared filesystem you can set '{OLMO_SHARED_FS_ENV_VAR}=1' instead."
         )