+++ src/olmo_core/internal/experiment.py
@@ -123,6 +123,8 @@ def build_common_components(
     if "jupiter" in cluster:
         root_dir = "/weka/oe-training-default/ai2-llm"
         weka_buckets.append(BeakerWekaBucket("oe-training-default", "/weka/oe-training-default"))
+    elif "augusta" in cluster:
+        root_dir = "gs://ai2-llm"
 
     beaker_user = (Beaker.from_env().account.whoami().name).upper()
     cmd_to_launch = SubCmd.train
@@ -181,7 +183,7 @@ def build_common_components(
             name=VSLCurriculumType.grow_p2, num_cycles=8, balanced=False
         ),
         work_dir=(
-            None
+            "./dataset-cache"
             if is_url(root_dir)
             else f"{root_dir}/checkpoints/{beaker_user.lower()}/dataset-cache"
         ),
@@ -205,7 +207,7 @@ def build_common_components(
                 sequence_length=dataset_config.effective_sequence_length,
                 tokenizer=tokenizer_config,
                 work_dir=(
-                    None
+                    "./dataset-cache"
                     if is_url(root_dir)
                     else f"{root_dir}/checkpoints/{beaker_user.lower()}/dataset-cache"
                 ),