+++ src/olmo_core/launch/beaker.py
@@ -316,6 +316,7 @@ class BeakerLaunchConfig(Config):
         entrypoint_script = [
             "#!/usr/bin/env bash",
             "set -exuo pipefail",
+            "[[ -d /var/lib/tcpxo/lib64 ]] && export LD_LIBRARY_PATH=/var/lib/tcpxo/lib64:$LD_LIBRARY_PATH",
             "mkdir -p /olmo-core-runtime",
             "cd /olmo-core-runtime",
             *self.setup_steps,